create result namespace
