--- lib/traffic_source_parser/parser/campaign_parser/campaign.rb
@@ -1,15 +1,0 @@
-require 'ostruct'
-
-module TrafficSourceParser
-  module Parser
-    module CampaignParser
-      class Campaign < OpenStruct
-
-        def initialize(campaign_hash)
-          super(campaign_hash)
-        end
-
-      end
-    end
-  end
-end
