--- lib/traffic_source_parser/parser/referrer_parser/generic.rb
@@ -1,16 +1,0 @@
-require 'ostruct'
-
-module TrafficSourceParser
-  module Parser
-    module ReferrerParser
-      class Generic < OpenStruct
-
-        def initialize(referrer)
-          super(source: referrer)
-        end
-
-      end
-    end
-
-  end
-end
