--- lib/traffic_source_parser/parser/referrer_parser/social.rb
@@ -1,15 +1,0 @@
-require 'ostruct'
-
-module TrafficSourceParser
-  module Parser
-    module ReferrerParser
-      class Social < OpenStruct
-
-        def initialize(source, referrer = nil)
-          super(source: source)
-        end
-
-      end
-    end
-  end
-end
