--- lib/traffic_source_parser/parser/utmz_parser/utmz.rb
@@ -1,15 +1,0 @@
-require 'ostruct'
-
-module TrafficSourceParser
-  module Parser
-    module UtmzParser
-      class Utmz < OpenStruct
-
-        def initialize(utmz_hash)
-          super(utmz_hash)
-        end
-
-      end
-    end
-  end
-end
