+++ lib/traffic_source_parser/result/campaign.rb
@@ -1,0 +1,13 @@
+require 'ostruct'
+
+module TrafficSourceParser
+  module Result
+    class Campaign < OpenStruct
+
+      def initialize(campaign_hash)
+        super(campaign_hash)
+      end
+
+    end
+  end
+end
