+++ lib/traffic_source_parser/result/generic.rb
@@ -1,0 +1,14 @@
+require 'ostruct'
+
+module TrafficSourceParser
+  module Result
+
+    class Generic < OpenStruct
+
+      def initialize(ref)
+        super(source: ref)
+      end
+
+    end
+  end
+end
