+++ lib/traffic_source_parser/result/search.rb
@@ -1,0 +1,14 @@
+require 'ostruct'
+
+module TrafficSourceParser
+  module Result
+
+    class Search < OpenStruct
+
+      def initialize(search_hash)
+        super(search_hash)
+      end
+
+    end
+  end
+end
