+++ lib/traffic_source_parser/result/social.rb
@@ -1,0 +1,14 @@
+require 'ostruct'
+
+module TrafficSourceParser
+  module Result
+
+    class Social < OpenStruct
+
+      def initialize(social_hash)
+        super(social_hash)
+      end
+
+    end
+  end
+end
