+++ lib/traffic_source_parser/result/utmz.rb
@@ -1,0 +1,13 @@
+require 'ostruct'
+
+module TrafficSourceParser
+  module Result
+    class Utmz < OpenStruct
+
+      def initialize(utmz_hash)
+        super(utmz_hash)
+      end
+
+    end
+  end
+end
